Move subdirectories into a sedonalib directory. Also in a few places replaced the print statement with the print function imported from __future__ so it works with Python 2. Also replaced import .. with imports from the sedonalib package and filled out __init__.py to import package classes and functions.
