--- data/atomic_data/__init__.py~
@@ -1,2 +1,0 @@
-# This file is intentionally blank. It is necessary for
-# making tools/ a Python package
