--- plot/__init__.py
@@ -1,1 +1,0 @@
-# This file has been intentionally left empty.
