+++ sedonalib/analysis/lightcurve/lcfilt.py
