+++ sedonalib/data/lightcurve_filters/filter_lambda_eff.dat
