+++ sedonalib/data/lightcurve_filters/filter_list.txt
