+++ sedonalib/loaders/__init__.py
@@ -1,0 +1,2 @@
+from .file_reader import SimReader, HDF5Reader, choose_reader
+
