+++ sedonalib/loaders/file_reader.py
