+++ sedonalib/plot/__init__.py
@@ -1,0 +1,2 @@
+from .grid import plot_grid
+
